requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-heroku==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
whitenoise==6.2.0
